Split controller into provider + reconciler
Following the example of other new sources, this splits the generated
controller into a provider and reconciler and moves the rbac
annotations into a doc.go.

The generated `ReconcileGitHubEventSource` type was also renamed to
just `reconciler`.
